add failing tests, and add unfinished tag to old tests that must be redone now that user permission will be checked on all add,edit,delete functions

+++ MOSTregi/events/views.py
@@ -3,6 +3,8 @@ from django.shortcuts import render_to_response, redirect, render, get_object_or
 from django.http import Http404, HttpResponseRedirect, HttpResponse
 from django.template import loader, TemplateDoesNotExist
 from django.utils import timezone
+from django.views.generic.edit import CreateView
+from django.contrib.auth.decorators import login_required, permission_required
 
 # project imports
 from .models import BookingRequest
@@ -16,22 +18,16 @@ def list_all(request):
     events = BookingRequest.objects.all().order_by('-date_time_received')
     return render(request, 'events/event_list.html', {'events': events})
 
+# @login_required(login_url='/login/')
+# @permission_required('events.add_BookingRequest', raise_exception=True)
 def new(request):
     if request.method == "POST":
         form = BookingRequestForm(request.POST)
         if form.is_valid():
-            post = form.save(request.POST)
+            post = form.save(commit=False)
+            post.user = request.user
             post.save()
-            # this technically redundant.../
-            # form.save() should save the form data into a models
-            # but since form.save() also returns the instance of telephone
-            # model save, it doesn't hurt to just save it again, right?
-            #
-            # refactoring that when I feel like I actually get django...
-            #
             return redirect('events:show_detail', pk=post.pk)
-            # we wanted to use this... but it doesn't work!
-            # return redirect(reverse("events:confirm"), pk=post.pk)
         else:
             print("form not valid, form errors: %s, form is bound: %s" % (form.errors.as_data(), form.is_bound))
             data=request.POST.get('date_request')
@@ -40,7 +36,11 @@ def new(request):
         form = BookingRequestForm
     return render(request, 'events/new.html', {'form': form})
 
+# @login_required(login_url='/login/')
+# @permission_required('events.delete_BookingRequest', raise_exception=True)
 def delete(request, pk):
+    if not request.user.is_authenticated:
+        return render(request, 'error.html', {'error': "Permission Denied. Please log in before deleting."})
     try:
         event = BookingRequest.objects.get(pk=pk)
         deletetion_collector = event.delete()
@@ -50,6 +50,8 @@ def delete(request, pk):
     except BookingRequest.DoesNotExist:
         return render(request, 'error.html', {'error': "Form already deleted."})
 
+# @login_required(login_url='/login/')
+# @permission_required('events.edit_BookingRequest', raise_exception=True)
 def edit(request, pk):
     event = get_object_or_404(BookingRequest, pk=pk)
     if request.method == "POST":
